the build works. finally
